+++ src/nbl/video/CVulkanPhysicalDevice.cpp
@@ -625,7 +625,7 @@ std::unique_ptr<CVulkanPhysicalDevice> CVulkanPhysicalDevice::create(core::smart
             
         properties.limits.dispatchBase = true;
         properties.limits.allowCommandBufferQueryCopies = true; // TODO: REDO WE NOW SUPPORT PERF QUERIES always true in vk for all query types instead of PerformanceQuery which we don't support at the moment (have VkPhysicalDevicePerformanceQueryPropertiesKHR::allowCommandBufferQueryCopies in mind)
-        properties.limits.maxOptimallyResidentWorkgroupInvocations = hlsl::min(core::roundDownToPoT(properties.limits.maxComputeWorkGroupInvocations),512u);
+        properties.limits.maxOptimallyResidentWorkgroupInvocations = hlsl::min(core::roundDownToPoT(properties.limits.maxComputeWorkGroupInvocations),static_cast<uint16_t>(512));
             
         auto invocationsPerComputeUnit = getMaxInvocationsPerComputeUnitsFromDriverID(properties.driverID);
         if(isExtensionSupported(VK_NV_SHADER_SM_BUILTINS_EXTENSION_NAME))
